sync

+++ src/controller_model/distributionList.ts
@@ -15,9 +15,13 @@
  * along with Threema Web. If not, see <http://www.gnu.org/licenses/>.
  */
 
-import {ControllerModelMode} from '../types/enums';
+import {WebClientService} from '../services/webclient';
 
-export class DistributionListControllerModel implements threema.ControllerModel {
+// Type aliases
+import ControllerModelMode = threema.ControllerModelMode;
+
+export class DistributionListControllerModel implements threema.ControllerModel<threema.DistributionListReceiver> {
+    private logTag = '[DistributionListControllerModel]';
 
     private $log: ng.ILogService;
     private $translate: ng.translate.ITranslateService;
@@ -28,36 +32,42 @@ export class DistributionListControllerModel implements threema.ControllerModel 
     public isLoading = false;
 
     private addContactPlaceholder: string;
-    private distributionList: threema.DistributionListReceiver;
-    private webClientService: threema.WebClientService;
+    private distributionList: threema.DistributionListReceiver | null;
+    private webClientService: WebClientService;
     private mode: ControllerModelMode;
-    private onRemovedCallback: any;
+    private onRemovedCallback: threema.OnRemovedCallback;
 
     constructor($log: ng.ILogService, $translate: ng.translate.ITranslateService, $mdDialog: ng.material.IDialogService,
-                webClientService: threema.WebClientService,
+                webClientService: WebClientService,
                 mode: ControllerModelMode,
-                distributionList: threema.DistributionListReceiver = undefined) {
+                distributionList?: threema.DistributionListReceiver) {
         this.$log = $log;
         this.$translate = $translate;
         this.$mdDialog = $mdDialog;
 
-        this.distributionList = distributionList;
+        if (distributionList === undefined) {
+            if (mode !== ControllerModelMode.NEW) {
+                throw new Error('DistributionListControllerModel: Distribution list ' +
+                                'may not be undefined for mode ' + mode);
+            }
+        } else {
+            this.distributionList = distributionList;
+        }
         this.mode = mode;
         this.webClientService = webClientService;
         this.addContactPlaceholder = $translate.instant('messenger.DISTRIBUTION_LIST_SELECT_MEMBERS');
 
         switch (this.getMode()) {
             case ControllerModelMode.EDIT:
-                this.subject = $translate.instant('messenger.EDIT_RECEIVER', {
-                    receiverName: '@NAME@',
-                }).replace('@NAME@', this.distributionList.displayName);
-                this.name = this.distributionList.displayName;
-                this.members = this.distributionList.members;
+                this.subject = $translate.instant('messenger.EDIT_RECEIVER');
+                this.name = this.distributionList!.displayName;
+                this.members = this.distributionList!.members;
                 break;
 
             case ControllerModelMode.VIEW:
-                this.subject = this.distributionList.displayName;
-                this.members = this.distributionList.members;
+            case ControllerModelMode.CHAT:
+                this.subject = this.distributionList!.displayName;
+                this.members = this.distributionList!.members;
                 break;
 
             case ControllerModelMode.NEW:
@@ -66,11 +76,11 @@ export class DistributionListControllerModel implements threema.ControllerModel 
                 break;
 
             default:
-                $log.error('Invalid controller model mode: ', this.getMode());
+                $log.error(this.logTag, 'Invalid controller model mode: ', this.getMode());
         }
     }
 
-    public setOnRemoved(callback: any): void {
+    public setOnRemoved(callback: threema.OnRemovedCallback): void {
         this.onRemovedCallback = callback;
     }
 
@@ -80,8 +90,12 @@ export class DistributionListControllerModel implements threema.ControllerModel 
 
     public isValid(): boolean {
         return this.members.filter((identity: string) => {
-                return identity !== this.webClientService.getMyIdentity().identity;
-            }).length > 0;
+            return identity !== this.webClientService.me.id;
+        }).length > 0;
+    }
+
+    public canChat(): boolean {
+        return true;
     }
 
     public canEdit(): boolean {
@@ -89,9 +103,53 @@ export class DistributionListControllerModel implements threema.ControllerModel 
         return true;
     }
 
-    public delete(ev): void {
+    public canClean(): boolean {
+        return true;
+    }
 
-        let confirm = this.$mdDialog.confirm()
+    public clean(ev: any): any {
+        const confirm = this.$mdDialog.confirm()
+            .title(this.$translate.instant('messenger.DELETE_THREAD'))
+            .textContent(this.$translate.instant('messenger.DELETE_THREAD_MESSAGE', {count: 1}))
+            .targetEvent(ev)
+            .ok(this.$translate.instant('common.YES'))
+            .cancel(this.$translate.instant('common.CANCEL'));
+
+        this.$mdDialog.show(confirm).then(() => {
+            this.reallyClean();
+        }, () => {
+            this.$log.debug(this.logTag, 'Clean canceled');
+        });
+    }
+
+    private reallyClean(): any {
+        if (!this.distributionList) {
+            this.$log.error(this.logTag, 'reallyClean: Distribution list is null');
+            return;
+        }
+        if (!this.canClean()) {
+            this.$log.error(this.logTag, 'Not allowed to clean this distribution list');
+            return;
+        }
+
+        this.isLoading = true;
+        this.webClientService.cleanReceiverConversation(this.distributionList)
+            .then(() => {
+                this.isLoading = false;
+            })
+            .catch((error) => {
+                // TODO: Handle this properly / show an error message
+                this.$log.error(this.logTag, `Cleaning receiver conversation failed: ${error}`);
+                this.isLoading = false;
+            });
+    }
+
+    public canShowQr(): boolean {
+        return false;
+    }
+
+    public delete(ev): void {
+        const confirm = this.$mdDialog.confirm()
             .title(this.$translate.instant('messenger.DISTRIBUTION_LIST_DELETE'))
             .textContent(this.$translate.instant('messenger.DISTRIBUTION_LIST_DELETE_REALLY'))
             .targetEvent(ev)
@@ -100,25 +158,30 @@ export class DistributionListControllerModel implements threema.ControllerModel 
 
         this.$mdDialog.show(confirm).then(() => {
             this.reallyDelete();
-
-            if (this.onRemovedCallback) {
-                this.onRemovedCallback(this.distributionList.id);
-            }
         }, () => {
-            this.$log.debug('delete canceled');
+            this.$log.debug(this.logTag, 'Delete canceled');
         });
     }
 
     private reallyDelete(): void {
+        if (!this.distributionList) {
+            this.$log.error(this.logTag, 'reallyDelete: Distribution list is null');
+            return;
+        }
         if (!this.distributionList.access.canDelete) {
-            this.$log.error('cannot delete distribution list');
+            this.$log.error(this.logTag, 'Not allowed to delete this distribution list');
             return;
         }
 
         this.isLoading = true;
         this.webClientService.deleteDistributionList(this.distributionList).then(() => {
             this.isLoading = false;
-        }).catch(() => {
+            if (this.onRemovedCallback && this.distributionList !== null) {
+                this.onRemovedCallback(this.distributionList.id);
+            }
+        }).catch((error) => {
+            // TODO: Handle this properly / show an error message
+            this.$log.error(this.logTag, `Deleting distribution list failed: ${error}`);
             this.isLoading = false;
         });
     }
@@ -127,7 +190,7 @@ export class DistributionListControllerModel implements threema.ControllerModel 
         switch (this.getMode()) {
             case ControllerModelMode.EDIT:
                 return this.webClientService.modifyDistributionList(
-                    this.distributionList.id,
+                    this.distributionList!.id,
                     this.members,
                     this.name,
                 );
@@ -136,12 +199,16 @@ export class DistributionListControllerModel implements threema.ControllerModel 
                     this.members,
                     this.name);
             default:
-                this.$log.error('not allowed to save distribution list');
-
+                this.$log.error(this.logTag, 'Cannot save distribution list, invalid mode');
+                return Promise.reject('Cannot save distribution list, invalid mode');
         }
     }
 
     public onChangeMembers(identities: string[]): void {
         this.members = identities;
+    }
+
+    public getMembers(): string[] {
+        return this.members;
     }
 }
